tests/requirements.txt
# 테스트 의존성
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
locust>=2.15.0
python-dotenv>=1.0.0

# 유닛 테스트 추가 의존성 (backend requirements에서 일부 복제)
pyjwt>=2.8.0
bcrypt>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
resend>=0.8.0
supabase>=2.0.0
